oracle-ai-langgraph/requirements.txt
oci
python-dotenv
oracle-ads
oracledb
pandas

langchain
langchain-community
langgraph
langchain-openai

copilotkit